refactor EMFStorage, move loading and saving to SiDiffResourceSet
Instead of eLoad and eSave, a SiDiffResourceSet should be created for an
operation. SiDiffResourceSet has utility functions for saving and
loading resources and objects.

URI/IFile/File conversion functions have been clean up, extended and
renamed. The functions operating on strings are deprecated, as using
URIs is superior to string modifications.