workaround for bug in EMF XMLSaveImpl, resulting in corrupted edit rules
The usage of the lookup table cache and configuration cache may result
in a Lookup being created, which has ExtendedMetaData=null, but a
different save operation, which requires ExtendedMetaData, will then
reuse the Lookup instance and throw a NullPointerException in
XMLSaveImpl$Lookup.getDocumentRoot. This is circumvented by always
setting the ExtendedMetaData to the default singleton instance.
Disabling the cache would also work, but reduce performance.

Also add some logging for errors and warnings attached to the resource
during load/save, so they can be detected earlier.