JSON based ExtensionSerialization:
ExtensionSerialization now serializes extensions and their configuration
in JSON instead of using the proprietary StringListSerializer, which
produces a lot of overhead in the output due to multiple escaping of
special characters.

The old format can still be imported to preserve backwards compatibility
with existing configuration and ecore files which contain extension
configurations.